Adding EULA acceptance: Fixing UTs
