+++ src/core/tests/Test_ConfigurationFactory.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 #
 # Requires Python 2.7+
-import json
-import os
+
 import unittest
 from core.src.bootstrap.Bootstrapper import Bootstrapper
 from core.src.bootstrap.Constants import Constants
@@ -30,9 +29,6 @@ class TestConfigurationFactory(unittest.TestCase):
 
     def tearDown(self):
         self.runtime.stop()
-
-    def mock_read_with_retry_raise_exception(self):
-        raise Exception
 
     def test_get_prod_config_correctly(self):
         bootstrapper = Bootstrapper(self.argument_composer, capture_stdout=False)
@@ -62,101 +58,6 @@ class TestConfigurationFactory(unittest.TestCase):
 
         self.assertEqual(config['package_manager_name'], Constants.APT)
         self.assertEqual(config['config_env'], Constants.DEV)
-
-    def test_eula_acceptance_file_read_success(self):
-        self.runtime.stop()
-
-        # Accept EULA set to true
-        eula_settings = {
-            "AcceptEULAForAllPatches": True,
-            "AcceptedBy": "TestSetup",
-            "LastModified": "2023-08-29"
-        }
-        f = open(Constants.AzGPSPaths.EULA_SETTINGS, "w+")
-        f.write(json.dumps(eula_settings))
-        f.close()
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, True)
-        runtime.stop()
-
-        # Accept EULA set to false
-        eula_settings = {
-            "AcceptEULAForAllPatches": False,
-            "AcceptedBy": "TestSetup",
-            "LastModified": "2023-08-29"
-        }
-        f = open(Constants.AzGPSPaths.EULA_SETTINGS, "w+")
-        f.write(json.dumps(eula_settings))
-        f.close()
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, False)
-        runtime.stop()
-
-    def test_eula_acceptance_file_read_when_no_data_found(self):
-        self.runtime.stop()
-
-        # EULA file does not exist
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, False)
-        self.assertFalse(os.path.exists(Constants.AzGPSPaths.EULA_SETTINGS))
-        runtime.stop()
-
-        # EULA settings set to None
-        eula_settings = None
-        f = open(Constants.AzGPSPaths.EULA_SETTINGS, "w+")
-        f.write(json.dumps(eula_settings))
-        f.close()
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, False)
-        self.assertTrue(os.path.exists(Constants.AzGPSPaths.EULA_SETTINGS))
-        runtime.stop()
-
-        # AcceptEULAForAllPatches not set in config
-        eula_settings = {
-            "AcceptedBy": "TestSetup",
-            "LastModified": "2023-08-29"
-        }
-        f = open(Constants.AzGPSPaths.EULA_SETTINGS, "w+")
-        f.write(json.dumps(eula_settings))
-        f.close()
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, False)
-        self.assertTrue(os.path.exists(Constants.AzGPSPaths.EULA_SETTINGS))
-        runtime.stop()
-
-        # AcceptEULAForAllPatches not set to a boolean
-        eula_settings = {
-            "AcceptEULAForAllPatches": "test",
-            "AcceptedBy": "TestSetup",
-            "LastModified": "2023-08-29"
-        }
-        f = open(Constants.AzGPSPaths.EULA_SETTINGS, "w+")
-        f.write(json.dumps(eula_settings))
-        f.close()
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        execution_config = container.get('execution_config')
-        self.assertEqual(execution_config.accept_package_eula, False)
-        self.assertTrue(os.path.exists(Constants.AzGPSPaths.EULA_SETTINGS))
-        runtime.stop()
-
-        runtime = RuntimeCompositor(self.argument_composer, True, package_manager_name=Constants.APT)
-        container = runtime.container
-        self.backup_read_with_retry = runtime.env_layer.file_system.read_with_retry
-        runtime.env_layer.file_system.read_with_retry = self.mock_read_with_retry_raise_exception
-        self.assertTrue(os.path.exists(Constants.AzGPSPaths.EULA_SETTINGS))
-        self.assertRaises(Exception, container.get('execution_config'))
-        runtime.stop()
 
 
 if __name__ == '__main__':
